requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# FRMIL/models/dataloaders/model40_meta.py: 6
# FRMIL/models/dataloaders/wsi_datasets.py: 10
# FRMIL/wsi_tools/compute_feats.py: 11
# FRMIL/wsi_tools/loader.py: 21
# FRMIL/wsi_tools/sample_spot.py: 11
# FRMIL/wsi_tools/utils.py: 5
Pillow == 8.4.0

# FRMIL/wsi_tools/compute_feats.py: 1
# FRMIL/wsi_tools/create_lib.py: 1
# FRMIL/wsi_tools/sample_spot.py: 2
# FRMIL/wsi_tools/xml2mask.py: 2
PyYAML == 5.4.1

# FRMIL/models/dataloaders/model40.py: 2
# FRMIL/models/dataloaders/model40_meta.py: 4
h5py == 2.10.0

# FRMIL/wsi_tools/xml2mask.py: 11
histomicstk == 1.1.0

# FRMIL/models/dataloaders/mil_datasets.py: 22
# FRMIL/plots_cm16.py: 18,19,22
# FRMIL/wsi_tools/loader.py: 12
# FRMIL/wsi_tools/sample_spot.py: 10
# FRMIL/wsi_tools/xml2mask.py: 7
matplotlib == 3.2.1

# FRMIL/common/meter.py: 1
# FRMIL/common/utils.py: 8
# FRMIL/models/dataloaders/mil_datasets.py: 23
# FRMIL/models/dataloaders/model40.py: 1
# FRMIL/models/dataloaders/model40_meta.py: 3
# FRMIL/models/dataloaders/samplers.py: 2
# FRMIL/models/dataloaders/wsi_datasets.py: 6
# FRMIL/models/mil_ss.py: 4
# FRMIL/plots_cm16.py: 17
# FRMIL/train.py: 5
# FRMIL/train_simclr.py: 6
# FRMIL/wsi_tools/compute_feats.py: 10
# FRMIL/wsi_tools/loader.py: 9
# FRMIL/wsi_tools/sample_spot.py: 9
# FRMIL/wsi_tools/utils.py: 2,3
# FRMIL/wsi_tools/xml2mask.py: 6
numpy == 1.19.2

# FRMIL/plots_cm16.py: 18
# FRMIL/wsi_tools/loader.py: 11
# FRMIL/wsi_tools/sample_spot.py: 12
# FRMIL/wsi_tools/xml2mask.py: 6
opencv_python_headless == 4.5.5.62

# FRMIL/wsi_tools/loader.py: 10
# FRMIL/wsi_tools/sample_spot.py: 8
# FRMIL/wsi_tools/xml2mask.py: 5
openslide_python == 1.1.2

# FRMIL/models/dataloaders/mil_datasets.py: 21
# FRMIL/plots_cm16.py: 20
# FRMIL/wsi_tools/xml2mask.py: 4,9
pandas == 1.1.5

# FRMIL/wsi_tools/loader.py: 17,18,19,253
# FRMIL/wsi_tools/utils.py: 7,8
# FRMIL/wsi_tools/xml2mask.py: 8
scikit_image == 0.17.2

# FRMIL/common/meter.py: 4,5
# FRMIL/models/dataloaders/mil_datasets.py: 14,15,16,18
scikit_learn == 0.23.0

# FRMIL/models/dataloaders/mil_datasets.py: 17
scipy == 1.5.2

# FRMIL/plots_cm16.py: 17
seaborn == 0.11.2

# FRMIL/common/utils.py: 9
termcolor == 1.1.0

# FRMIL/common/meter.py: 2
# FRMIL/common/utils.py: 2,4
# FRMIL/models/dataloaders/mil_datasets.py: 12
# FRMIL/models/dataloaders/model40.py: 2
# FRMIL/models/dataloaders/model40_meta.py: 1,7
# FRMIL/models/dataloaders/samplers.py: 1
# FRMIL/models/dataloaders/wsi_datasets.py: 4,5,6,8
# FRMIL/models/mil_dsmil.py: 1,2,3
# FRMIL/models/mil_ss.py: 1,2,3
# FRMIL/models/resnet_simclr.py: 1,2,3
# FRMIL/plots_cm16.py: 3,4,5,7
# FRMIL/test.py: 3,4,5,7
# FRMIL/train.py: 5,6,7,9
# FRMIL/train_simclr.py: 6,7,8,10
# FRMIL/wsi_tools/compute_feats.py: 5,8
# FRMIL/wsi_tools/create_lib.py: 1
# FRMIL/wsi_tools/loader.py: 15
# FRMIL/wsi_tools/patch_gen.py: 6
# FRMIL/wsi_tools/sample_spot.py: 4
torch == 1.7.1

# FRMIL/models/dataloaders/mil_datasets.py: 11
# FRMIL/models/dataloaders/wsi_datasets.py: 7
# FRMIL/models/resnet_simclr.py: 4
# FRMIL/wsi_tools/compute_feats.py: 6,7
# FRMIL/wsi_tools/loader.py: 14
torchvision == 0.8.2

# FRMIL/models/dataloaders/mil_datasets.py: 19
# FRMIL/plots_cm16.py: 2
# FRMIL/test.py: 2
# FRMIL/train.py: 2
# FRMIL/train_simclr.py: 3
# FRMIL/wsi_tools/compute_feats.py: 12
# FRMIL/wsi_tools/loader.py: 8
# FRMIL/wsi_tools/patch_gen.py: 7
tqdm == 4.62.2

# FRMIL/common/utils.py: 3
# FRMIL/train.py: 4
# FRMIL/train_simclr.py: 5
wandb == 0.12.2
